Add InvoiceFactory with GenericInvoice
