+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/invoicing/BookKeeper.java
@@ -21,8 +21,8 @@ import pl.com.bottega.ecommerce.sharedkernel.Money;
 
 public class BookKeeper {
 
-    public Invoice issuance(ClientData client, List<RequestItem> items) {
-        Invoice invoice = new Invoice(Id.generate(), client);
+    public Invoice issuance(ClientData client, List<RequestItem> items, InvoiceFactory invoiceFactory) {
+        Invoice invoice = invoiceFactory.createGenericInvoice(client);
 
         for (RequestItem item : items) {
             Money net = item.getTotalCost();
