+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/invoicing/InvoiceFactory.java
@@ -1,0 +1,10 @@
+package pl.com.bottega.ecommerce.sales.domain.invoicing;
+
+import pl.com.bottega.ecommerce.canonicalmodel.publishedlanguage.ClientData;
+import pl.com.bottega.ecommerce.canonicalmodel.publishedlanguage.Id;
+
+public class InvoiceFactory {
+    public Invoice createGenericInvoice(ClientData client) {
+        return new GenericInvoice(Id.generate(), client);
+    }
+}
